Make database validate again

+++ database/migrations/2018_07_03_091314_create_application_metrics_table.php
@@ -16,8 +16,8 @@ class CreateApplicationMetricsTable extends Migration
         Schema::create('application_metrics', function (Blueprint $table) {
             $table->integer('app_id');
             $table->string('metric', 32);
-            $table->integer('value')->nullable();
-            $table->integer('value_prev')->nullable();
+            $table->double('value')->nullable();
+            $table->double('value_prev')->nullable();
             $table->unique(['app_id','metric'], 'application_metrics_app_id_metric_uindex');
         });
     }
